pyproject.toml
[project]
name = "oap"
version = "0.1.0"
description = "OA系统通知爬取与推送项目"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bs4>=0.0.2",
    "requests>=2.32.5",
    "psycopg[binary]>=3.2.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "flask-limiter>=3.5.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
